requirements.txt
dash>=2.16
dash-canvas>=0.1.0
numpy>=1.24
pandas>=2.0
plotly>=5.20
pydicom>=2.4
Pillow>=10.0
scipy>=1.11
scikit-image>=0.22
